+++ CHANGELOG.md
@@ -6,7 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 - Mongodb Enterprise - Delete default alerts on service creation
-
+- Adds context object to update request
+- Adapts Kubernetes Provisioning 
 
 
 ## [4.2.4] - 2018-09-06